requirements.txt
einops
faster-coco-eval
graphviz
hydra-core
lightning
loguru
numpy
opencv-python
Pillow
pycocotools
requests
rich
wandb
tabulate
ipywidgets